+++ lib/meta/attributes.dart
@@ -33,7 +33,6 @@ class Attribute extends Property {
   AttributeType get type => _type;
   set type(AttributeType attributeType) {
     _type = attributeType;
-    //if (attributeType != null && length == null) {
     if (attributeType != null) {
       length = attributeType.length;
     }