+++ CHANGELOG.md
@@ -4,7 +4,8 @@ based on [Semantic Versioning](http://semver.org/)
 
 *2.0.0* 2014-04-01
 
-+ return null if getStringFromAttribute is null
++ meta: add label and labels in concept; label in property; length in attribute 
++ return null, and not "null", if getStringFromAttribute is null
 + add a list of not increment attributes in concept
 + add a list of identifier attributes in concept
 + add a list of not identifier attributes in concept