+++ CHANGELOG.md
@@ -5,8 +5,8 @@ based on [Semantic Versioning](http://semver.org/)
 *2.0.0* 2014-04-01
 
 + add Telephone, Name, Description and Money attribute types
-+ derive default labels from code and codes
-+ derive default attribute length from type length
++ derive default labels, if null, from code and codes
++ derive default attribute length, if null, from type length
 + meta: set length of Uri type, Email type
 + meta: add label and labels in concept; label in property; length in attribute
 + add getStringOrNullFromAttribute to ConceptEntity API: